refactor(dto): dto 패키지의 위치를 racingcar 하위로 이동
