+++ docs/README.md
@@ -180,6 +180,8 @@
         - A가 B를 의존한다는 것은 A가 B를 사용하고 있어서 B가 변경되면 A도 변경될 수도 있다는 의미로 이해했습니다. 이 부분을 알고나니 왜 domain 부분에서 View의 세부사항을 알필요가 없는지
           이해했습니다.
         - 이 요구사항을 만족하기위해 domain의 객체에서 toString()을 view로 보내지 않고 DTO를 이용하여 view로 필요한 데이터를 보내주게 되었습니다.
+    - DTO의 위치를 racingcar 하위로 이동함.
+        - DTO는 View, Domain모두에서 사용하기 때문에 domain 패키지 하위에 있을 필요가 없다고 생각했다.
 - [x] DTO들이 필드를 하나씩만 가지고 있는데, DTO를 쓸 필요가 있을까요?
     - 이전 코드를 작성할 당시에는 DTO에 대해 제대로 이해하지 않고 사용하면서 조금은 무의식적으로 사용했던것 같습니다. 공부를 하고난 뒤 이 부분을 보면서 DTO가 필요한 이유를 고민해봤습니다.
     - DTO가 필드를 하나만 가지고 있더라도 DTO를 이용하여 값을 전달하면 domain에서 데이터를 가져오기 전까지 변조가 되지 않는 장점이 있을 것 같습니다. Contoller에서 View와
