+++ src/main/java/racingcar/dto/RequestAttemptDto.java
@@ -1,4 +1,4 @@
-package racingcar.domain.dto;
+package racingcar.dto;
 
 import racingcar.domain.vo.Attempt;
