+++ src/main/java/racingcar/dto/RequestCarsDto.java
@@ -1,4 +1,4 @@
-package racingcar.domain.dto;
+package racingcar.dto;
 
 import racingcar.domain.vo.Cars;
