+++ src/main/java/racingcar/view/InputView.java
@@ -2,8 +2,8 @@ package racingcar.view;
 
 import java.util.NoSuchElementException;
 import java.util.Scanner;
-import racingcar.domain.dto.RequestAttemptDto;
-import racingcar.domain.dto.RequestCarsDto;
+import racingcar.dto.RequestAttemptDto;
+import racingcar.dto.RequestCarsDto;
 import racingcar.domain.vo.Attempt;
 import racingcar.domain.vo.Cars;
 import racingcar.view.enums.ViewErrorMessage;
